backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
apscheduler==3.10.4
beautifulsoup4==4.12.3
lxml==5.3.0
pytest==8.3.4
pytest-asyncio==0.24.0
